ingestion/requirements.txt
apache-beam[gcp]==2.72.0
requests==2.33.1
google-cloud-secret-manager==2.27.0